\name{extractPointVgm}
\alias{extractPointVgm}
\title{
Extract point-scale variogram from deconvoluted ataKrigVgm.
}
\description{
Extract point-scale variogram from deconvoluted ataKrigVgm.
}
\usage{
extractPointVgm(g)
}
\arguments{
  \item{g}{
deconvoluted ataKrigVgm object.
}
}
\value{
a list of gstat vgm model.
}
