\name{rbindDiscreteArea}
\alias{rbindDiscreteArea}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Combine two discrete areas.
}
\description{
Combine two discrete areas.
}
\usage{
rbindDiscreteArea(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x, y}{
discretized area, list(areaValues, discretePoints).
}
}
\value{
discretized area, list(areaValues, discretePoints).
}
