\name{subsetDiscreteArea}
\alias{subsetDiscreteArea}
\title{
Select discrete area according to area id.
}
\description{
Select discrete area according to area id.
}
\usage{
subsetDiscreteArea(x, selAreaId, revSel = FALSE)
}
\arguments{
  \item{x}{
a discreteArea object: list(areaValues, discretePoints).
}
  \item{selAreaId}{
area id to select.
}
  \item{revSel}{
reverse select or not.
}
}
%%\details{}
\value{
a discreteArea object: list(areaValues, discretePoints).
}
